Tidy preference test specs.
